Rename BuildConfig => Config
