+++ src/builders/util.js
@@ -2,7 +2,7 @@
  * @flow
  */
 
-import type {BuildSpec, BuildConfig, BuildEnvironment} from '../types';
+import type {BuildSpec, Config, BuildEnvironment} from '../types';
 
 import * as child from 'child_process';
 import * as fs from '../lib/fs';
@@ -15,7 +15,7 @@ type ConfigSpec = {
 
 export function renderSandboxSbConfig(
   spec: BuildSpec,
-  config: BuildConfig,
+  config: Config,
   sandboxSpec?: ConfigSpec = {},
 ): string {
   const subpathList = pathList =>
